# More efficient Merge implementation.